Move table data to separate folder.
